Remove underscore syntax matching any symbol
Underscore syntax:

For instance, a tree `(expr (id 'a') (add '+') (_))` matches both source codes `a + b * c` and `a - b / c`.

This feature is helpful because it allows you to emphasize the main points of the test by ignoring nodes of no interest.
However, we will remove the feature for the time being to reconsider the grammar.