Set token positions for the EOF symbol
vartan intentionally had not set positional information for the EOF symbol in a token because the corresponding character does not exist in the source code.
However, to be able to include the positional information in a syntax error message when a syntax error occurs due to the detection of the EOF symbol during parsing,
this commit sets the position next to the last character in the source code as the position of the EOF symbol.